+++ src/core/course_hanlder.py
@@ -5,23 +5,24 @@ from app import db
 class CourseHanlder:
     @classmethod
     def get_info(cls, is_active):
-        info_list = list()
+        result_list = list()
         if is_active:
-            info_ = db.session.query(Course.is_active == is_active).first()
+            info_ = db.session.query(Course).filter(Course.is_active == is_active).all()
         else:
-            info_ = Course.query.all()
+            info_ = db.session.query(Course).all()
         for info in info_:
             results = {
                 'name': info.name,
                 'classroom_id': info.classroom_id,
-                'is_active': info.is_active,
-                'open_time': info.open_time,
+                'open_datetime': info.open_time.strftime("%Y-%m-%d %H:%M:%S"),
             }
-            info_list.append(results)
-        return info_list
+            result_list.append(results)
+        return result_list
 
     @classmethod
     def add_info(cls, name, classroom_id, is_active, open_time):
+        if not classroom_id:
+            raise ValueError('Classroom id not exist')
         info = Course(name=name,
                       classroom_id=classroom_id,
                       is_active=is_active,
@@ -34,6 +35,8 @@ class CourseHanlder:
     @classmethod
     def del_info(cls, name):
         info = db.session.query(Course).filter(Course.name == name).first()
+        if not info:
+            raise ValueError('Name not exist')
         db.session.delete(info)
         db.session.commit()
         return {'success': True}
@@ -41,19 +44,17 @@ class CourseHanlder:
     @classmethod
     def update_info(cls, name, classroom_id, is_active, open_time):
         info = db.session.query(Course).filter(Course.name == name).first()
-        if classroom_id:
-            Course.classroom_id = classroom_id
-        if is_active:
-            Course.is_active = is_active
-        if open_time:
-            Course.open_time = open_time
+        if not info:
+            raise ValueError('Name not exist')
+        Course.classroom_id = classroom_id
+        Course.is_active = is_active
+        Course.open_time = open_time
         db.session.add(info)
-        db.session.commit()
         results = {
             'name': info.name,
             'gender': info.gender,
             'phone_number': info.phone_number,
         }
+        db.session.commit()
         return results
-
