+++ src/core/emergency_hanlder.py
@@ -4,12 +4,12 @@ from app import db
 
 class EmergencyHandler:
     @classmethod
-    def get_all_info(cls, student_id):
+    def get_info(cls, student_id):
+        result_list = list()
         if student_id:
             emergency_contact = db.session.query(EmergencyContact).filter(EmergencyContact.student_id == student_id).first()
         else:
-            emergency_contact = db.session.query(EmergencyContact.name, EmergencyContact.student_id, EmergencyContact.relationship_to_client, EmergencyContact.phone_number).all()
-        result_list = list()
+            emergency_contact = db.session.query(EmergencyContact).all()
         for emergency in emergency_contact:
             result = {
                 'name': emergency.name,
@@ -21,7 +21,9 @@ class EmergencyHandler:
         return result_list
 
     @classmethod
-    def add_emergency_info(cls, name, student_id, relationship_to_client, phone_number):
+    def add_info(cls, name, student_id, relationship_to_client, phone_number):
+        if not student_id:
+            raise ValueError('Student id not exist')
         emergency = EmergencyContact(
             name=name,
             student_id=student_id,
@@ -33,21 +35,22 @@ class EmergencyHandler:
         return {'success': True}
 
     @classmethod
-    def delete_emergency_info(cls, student_id):
+    def delete_info(cls, student_id):
         emergency = db.session.query(EmergencyContact).filter(EmergencyContact.student_id == student_id).first()
+        if not student_id:
+            raise ValueError('Student id not exist')
         db.session.delete(emergency)
         db.session.commit()
         return {'success': True}
 
     @classmethod
-    def update_emergency_info(cls, name, phone_number, relationship):
+    def update_info(cls, name, phone_number, relationship):
         user = db.session.query(EmergencyContact).filter(EmergencyContact.name == name).first()
-        if phone_number:
-            user.phone_number = phone_number
-        if relationship:
-            user.relationship_to_client = relationship
+        if not name:
+            raise ValueError('Name not exist')
+        user.phone_number = phone_number
+        user.relationship_to_client = relationship
         db.session.add(user)
-        db.session.commit()
         results = {
             'id': user.id,
             'name': user.name,
@@ -55,4 +58,5 @@ class EmergencyHandler:
             'relationship_to_client': user.relationship_to_client,
             'phone_number': user.phone_number,
         }
+        db.session.commit()
         return results
