+++ src/core/school_hanlder.py
@@ -23,7 +23,6 @@ class SchoolHandler:
             user_list.append(results)
         return user_list
 
-
     @classmethod
     def add_info(cls, name, gender, grade, phone_number):
         student = Student(
@@ -32,6 +31,8 @@ class SchoolHandler:
             grade=grade,
             phone_number=phone_number,
         )
+        if not student:
+            raise ValueError('Wrong format')
         db.session.add(student)
         db.session.commit()
         return {'success': True}
@@ -39,6 +40,8 @@ class SchoolHandler:
     @classmethod
     def delete_info(cls, student_name):
         student = db.session.query(Student).filter(Student.name == student_name).first()
+        if not student:
+            raise ValueError('Student name not exist')
         db.session.delete(student)
         db.session.commit()
         return {'success': True}
@@ -46,10 +49,10 @@ class SchoolHandler:
     @classmethod
     def update_info(cls, student_name, phone_number):
         user = db.session.query(Student).filter(Student.name == student_name).first()
-        if phone_number:
-            user.phone_number = phone_number
+        if not user:
+            raise ValueError('Student name not exist')
+        user.phone_number = phone_number
         db.session.add(user)
-        db.session.commit()
         results = {
             'id': user.id,
             'name': user.name,
@@ -59,5 +62,6 @@ class SchoolHandler:
             'create_datetime': user.create_datetime.strftime("%Y-%m-%d %H:%M:%S"),
             'update_datetime': user.update_datetime.strftime("%Y-%m-%d %H:%M:%S"),
         }
+        db.session.commit()
         return results
