+++ src/core/teacher_hanlder.py
@@ -9,7 +9,7 @@ class TeacherHanlder:
         if name:
             user_ = db.session.query(Teacher).filter(Teacher.name == name).first()
         else:
-            user_ = db.session.query(Teacher.name, Teacher.gender, Teacher.phone_number).all()
+            user_ = db.session.query(Teacher).all()
         for user in user_:
             result = {
                 'name': user.name,
@@ -18,8 +18,6 @@ class TeacherHanlder:
             }
             result_list.append(result)
         return result_list
-
-
 
     @classmethod
     def add_info(cls, name, gender, phone_number):
@@ -35,6 +33,8 @@ class TeacherHanlder:
     @classmethod
     def delete_info(cls, name):
         user = db.session.query(Teacher).filter(Teacher.name == name).first()
+        if not user:
+            raise ValueError('User not exist')
         db.session.delete(user)
         db.session.commit()
         return {'success': True}
@@ -42,17 +42,18 @@ class TeacherHanlder:
     @classmethod
     def update_info(cls, name, phone_number, gender):
         user = db.session.query(Teacher).filter(Teacher.name == name).first()
-        if phone_number:
-            user.phone_number = phone_number
-        if gender:
-            user.gender = gender
+        if not user:
+            raise ValueError('Teacher name not exist')
+        user.phone_number = phone_number
+        user.gender = gender
         db.session.add(user)
-        db.session.commit()
         results = {
             'name': user.name,
             'gender': user.gender,
             'phone_number': user.phone_number,
         }
+        db.session.commit()
+
         return results
